separate createProject page to three pages
